refactor: extrac my collection sticky header to a separate file
