+++ src/app/Scenes/MyCollection/Screens/Artwork/Components/ArtworkInsights/MyCollectionArtworkArtistAuctionResults.tsx
@@ -8,10 +8,10 @@ import {
 import { SectionTitle } from "app/Components/SectionTitle"
 import { navigate } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
+import { useScreenDimensions } from "app/utils/hooks"
 import { FlatList } from "react-native"
 import { graphql, useFragment } from "react-relay"
 import { useTracking } from "react-tracking"
-import { useScreenDimensions } from "app/utils/hooks"
 
 interface MyCollectionArtworkArtistAuctionResultsProps {
   artwork: MyCollectionArtworkArtistAuctionResults_artwork$key
@@ -26,6 +26,8 @@ export const MyCollectionArtworkArtistAuctionResults: React.FC<
 
   const artist = artwork.artist
   const auctionResults = extractNodes(artist?.auctionResultsConnection)
+
+  const { width } = useScreenDimensions()
 
   if (!auctionResults.length) {
     return null
@@ -57,7 +59,7 @@ export const MyCollectionArtworkArtistAuctionResults: React.FC<
           </>
         )}
         ItemSeparatorComponent={AuctionResultListSeparator}
-        style={{ width: useScreenDimensions().width, left: -20 }}
+        style={{ width, left: -20 }}
       />
     </Flex>
   )
